Swich to use the ESO operator
This change drops the use of the ESO helm subchart which we have to
patch and switches to the use of the ESO operator. This means
that you *must* add ESO to your subscriptions in your values files:

    subscriptions:
      eso:
        channel: alpha
        name: external-secrets-operator
        namespace: golang-external-secrets
        source: community-operators

The reason we use sync-waves in the golang-external-secrets service
account token and in the clustersecretstore is because they need to
be applied *after* the OperatorConfig is created. That is because the
OperatorConfig creates the service account and the secret to add a token
to the service account won't get applied correctly, unless the service
account is present and this confuses argo greatly.

Tested correctly on a full MCG hub and spoke.